requirements.txt
asgiref==3.3.1
cffi==1.14.6
Django==3.1.7
django-bootstrap3==15.0.0
djangorestframework==3.12.4
Pillow==8.1.2
pycparser==2.20
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
